requirements.txt
# Core Framework
flask==2.3.3
requests==2.31.0

# Cryptography (Phase 1)
cryptography==41.0.4
charm-crypto==0.50
pyclamd==0.4.0

# Azure Integration (Phase 2)
azure-storage-blob==12.19.0

# Configuration Management (Phase 2)
python-dotenv==1.0.0
pydantic==2.4.2

# Async File Operations (Phase 2)
aiofiles==23.2.1
